update add car and edit car

+++ BackEnd/controllers/car/types.go
@@ -7,10 +7,10 @@ type CarBody struct {
 	BrandCar        string  `binding:"required"`
 	ColorCar        string  `binding:"required"`
 	DescriptionCar  string
-	KilometersCar   int    `binding:"required"`
-	SeatsCar        int    `binding:"required"`
-	TransmissionCar string `binding:"required"`
-	AddressOwner    string `binding:"required"`
+	KilometersCar   float64 `binding:"required"`
+	SeatsCar        int     `binding:"required"`
+	TransmissionCar string  `binding:"required"`
+	AddressOwner    string  `binding:"required"`
 	ImagesCar       string
 	StatusCar       string `binding:"required"`
 }
