+++ FrontEnd/lib/domain/model/car.dart
@@ -14,7 +14,7 @@ class Car with _$Car {
     @Default('') String brandCar,
     @Default('') String colorCar,
     @Default('') String descriptionCar,
-    @Default(0) int kilometersCar,
+    @Default(0) double kilometersCar,
     @Default(0) int seatsCar,
     @Default('') String transmissionCar,
     @Default('') String addressCar,
